Atualizar Vendedora

+++ src/vendedora/vendedora.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateVendaDto } from 'src/venda/dto/createVendaDto.dto';
 import { CreateVendedoraDto } from './dto/createVendedoraDto.dto';
@@ -23,8 +23,19 @@ export class VendedoraService {
         return await this.repository.insertVendedora(createVendedoraDto)       
     }
 
-    async removeVenda(id: number): Promise<void>{
-        this.repository.delete(id)
-        return
+    async removeVenda(id: number): Promise<number>{
+        if(id == 1)
+            throw new BadRequestException("Não é possível deletar vendedora principal")
+
+        console.log(id)
+
+        
+        await this.repository.delete(id)
+        return id
     }
+
+    async updateVendedora(id, createVendedoraDto: CreateVendedoraDto): Promise<Vendedora>{
+        return await this.repository.updateVendedora(id, createVendedoraDto)
+    }
+
 }
